Extract newSources
