+++ src/actions/sources/index.js
@@ -1,2 +1,11 @@
+/* This Source Code Form is subject to the terms of the Mozilla Public
+ * License, v. 2.0. If a copy of the MPL was not distributed with this
+ * file, You can obtain one at <http://mozilla.org/MPL/2.0/>. */
+
+// @flow
+
 export { loadSourceText } from "./loadSourceText";
 export { togglePrettyPrint } from "./prettyPrint";
+export { newSource, newSources } from "./newSources";
+export { toggleBlackBox } from "./blackbox";
+export * from "./tabs";
